:recycle: Extracting end-to-end test code into context class
