+++ src/Tests/Console/EndToEnd/Happy_path.cs
@@ -1,58 +1,25 @@
-﻿using System;
-using System.Diagnostics;
-using System.IO;
-using NUnit.Framework;
+﻿using NUnit.Framework;
 
 namespace Fettle.Tests.Console.EndToEnd
 {
     [TestFixture]
-    public class Happy_path
+    public class Happy_path : Contexts.EndToEnd
     {
+        public Happy_path()
+        {
+            When_running_real_fettle_console_executable(configFilename: "fettle.config.yml");
+        }
+
         [Test]
         public void Console_executable_runs_and_outputs_info_about_surviving_mutants()
         {
-            var baseDir = TestContext.CurrentContext.TestDirectory;
-            var configFilePath = Path.Combine(baseDir, "Console", "fettle.config.yml");
-            ModifyConfigFile(configFilePath);
-
-            var fettleProcess = new Process
-            {
-                StartInfo = new ProcessStartInfo
-                {
-                    CreateNoWindow = true,
-                    WindowStyle = ProcessWindowStyle.Hidden,
-                    FileName = Path.Combine(baseDir, "Fettle.Console.exe"),
-                    Arguments = $"--quiet --config {configFilePath}",
-                    UseShellExecute = false,
-                    RedirectStandardOutput = true,
-                    RedirectStandardError = true
-                }
-            };
-
-            var stopwatch = Stopwatch.StartNew();
-
-            fettleProcess.Start();
-            fettleProcess.WaitForExit((int)TimeSpan.FromMinutes(1).TotalMilliseconds);
-
-            stopwatch.Stop();
-
-            var stdOut = fettleProcess.StandardOutput.ReadToEnd();
-            var stdErr = fettleProcess.StandardError.ReadToEnd();
-            
             Assert.Multiple(() =>
             {
-                Assert.That(fettleProcess.ExitCode, Is.EqualTo(1));
-                Assert.That(stdErr, Is.Empty);
-                Assert.That(stdOut, Does.Contain("mutant(s) survived!"));
-                Assert.That(stdOut, Does.Contain("PartiallyTestedNumberComparison.cs:7"));    
+                Assert.That(ExitCode, Is.EqualTo(1));
+                Assert.That(StandardError, Is.Empty);
+                Assert.That(StandardOutput, Does.Contain("mutant(s) survived!"));
+                Assert.That(StandardOutput, Does.Contain("PartiallyTestedNumberComparison.cs:7"));    
             });
-        }
-
-        private static void ModifyConfigFile(string configFilePath)
-        {
-            var originalConfigFileContents = File.ReadAllText(configFilePath);
-            File.WriteAllText(configFilePath, 
-                string.Format(originalConfigFileContents, BuildConfig.AsString));
         }
     }
 }
