+++ src/Tests/Tests.csproj
@@ -54,6 +54,7 @@
     <Compile Include="BuildConfig.cs" />
     <Compile Include="Console\Configuration.cs" />
     <Compile Include="Console\Contexts\Default.cs" />
+    <Compile Include="Console\Contexts\EndToEnd.cs" />
     <Compile Include="Console\Contexts\SourceControlIntegration.cs" />
     <Compile Include="Console\Contexts\SpyOutputWriter.cs" />
     <Compile Include="Console\CoverageAnalysis\Coverage_analysis_disabled.cs" />
